Add the basic Relational type class and some supporting types.
I've added the Relational.Class module, which contains the definition
of the 'Relational' type class, and types for selection conditions.
The conditions could have just been boolean functions, but I'd like
to be able to support a generic query optimizer, so I've given
conditions some structure.

The Relational class is multi-parameter:

 class (Ord n, Ord d) => Relational n d r | r -> n, r -> d

where 'n' is the type of attribute names, 'd' is the type of
attribute values, and 'r' is the type of relations. Currently,
the class contains methods for seven basic operations from
the relational algebra:

 - union
 - difference
 - attribute renaming
 - projection
 - selection
 - theta join
 - cartesian product

The theta join and the cartesian product are not independent; each
can be implemented given the other, so each method has a default
implementation. Since each of these operations imposes constraints
on the sets of attribute names for their argument relations,
they each yield results in an arbitrary error monad.

In addition to the relational algebra operations, there are
operations for extracting the signature and data of a relation,
and for forming a relation given a signature and a list of tuples.

I'd like to make a few more additions to the Relational class:

1. improving the Haddock documentation

2. adding methods for several non-primitive operations from the
   relational algebra, like outer joins. These can be given
   default implementations.

3. adding an (optional) method for the transitive closure of
   binary relations

(1) is probably the next task, but after that I'd like to write
a simple implementation for the class and some QuickCheck tests.