Rename a couple of functions in Relational.Naive.
I've used the convention of prefixing functions that operate
on Relations with 'rel', except with 'empty' and
'unsafeAddTuple'. To be consistent, I've renamed those
functions to 'relEmpty' and 'relUnsafeAddTuple'.

The 'rel' naming convention is arguably unnecessary,
especially if I split 'Relation' into its own module.
However, I'll keep for now in order to distinguish
Relation-specific functions from methods of the Relational
type class.