Replace 'CondTupleTest' with 'CondCall' in Condition.
Previously we supported an arbitrary tuple test with

 CondTupleTest ((n -> m d) -> m Bool)

But this interferes with query optimization. The optimizer
wouldn't know which attribute names the test function
calls, and so must assume that it will use all names.

Instead, take the approach used in Expression, and replace
this variant with

 CondCall ([d] -> m Bool) [Expression n d m]

In this variant, the test function cannot hide references
to attributes, and we can determine from the Expressions
which attributes are used.