Import Signature qualified in the Relational.Naive.
Signature has some functions that are most naturally named
so they clash with functions in the Prelude or in the
Relational module itself ('union' in particular). At the
moment, I've worked around that by giving what should be
'null' and 'union' the names 'isEmpty' and 'relUnion'.

Prepare to rename the functions by using a qualified import
in Relational.Naive. The renames come in the next commit.