Add a (rather indirect) test for 'make'.
I've added a test for 'make' that checks that when we make
a relational value out of given names and tuples, we get
equivalent names and tuples out of the resulting relation.

The test is indirect because I couldn't actually invoke
Relational.Class.make inside the test method, because there
is no way to deduce the type of the created relation.

This makes me think that perhaps 'make' does not belong
in the Relational class. Instead, creation methods need
to be specific to each type. I'm not positive about this
though, so I won't change the type class yet.

Also, for the new properties in RelationalProps, I've added
two closely related 'props':

 prop_makeSigAndTuples

and

 propM_makeSigAndTuples

The second is more fundamental, and returns a Bool in an
error monad; the first just forces the monad to be
(Either String) with a use of 'noErr'. I realized that
the separation is important because we cannot use the
(Either String) property functions with implementations
that must use a closed monad -- generally the IO monad,
or a transformed IO monad. I'll have to make a similar
split for the other 'prop_X' functions.