Added the SubList module below test.
I can't find a QuickCheck generator for creating sub-lists of
a given list, so here's one. At heart is a generator for
subsequences of [0..n], although that generator isn't
exposed.

I don't think the distribution of generated sub-sequences
is exactly uniform; it may tend to bias toward sequences
which have elements bunched toward the left. I haven't
proven this though, and haven't run tests to show a
non-uniform distribution. It seems to generate a reasonable
variety of sub-sequences for the purposes here.