Test that excluding single attributes commutes.
Terminology: we exclude attribute 'ak' from signature '{a1,...,an}'
by projecting all the attributes except 'ak'. I don't actually use
this terminology in the Relational class. Perhaps a more familiar
name would be 'drop'.

Non-overlapping exclusions should commute. This commit adds a test
for this. Treat this as the first test for 'project'.