Added generators for the types in Relational.Condition.
This includes generators for Expression, RelOp, and Condition.
RelOp is Arbitrary, but we create generators for Expression
and Condition from name generators. For tests for 'select'
and 'project', we'll need to make sure we generate expressions
and conditions that use only names from a given signature.

At the moment we can't use 'sample' with these generators
because there are no Show instances for Expression and
Condition. I'll probably add instances that show functions
as '<function>', just so I can get an idea of what the
generated data will look like.

The next step is to write a generator that takes an input
Condition and generates tuples that satisfy it. This may
require a more refined generator for Conditions, since
we want to generate Conditions that are satisfiable.