Added a generator for satisfiable Conditions.
For testing 'select' and 'join', we'll need to be able to
generate satisfiable conditions. On value types which are
Bounded, Enum, and Ord, I've added a brute force generator
that generates random conditions until it finds one
that can be satisfied. For small signatures and small
conditions I think this will perform OK, but I'll see in
the next few commits.