Rework the satisfiable condition generator a bit.
The old 'satisfiableCondition' returned the condition but not
the satisfying tuples. The satisfying tuples are of interest
for generating test relations for selection and join, so I've
reworked the definition, and the helper functions. I've also
added

 unsatisfiableCondition

  for generating conditions that have no satisfying tuples,

and

 conditionAndSatisfyingTuples

  for generating arbitrary conditions and the possibly empty
  lists of tuples that satisfy them

I should really define a newtype for a tuple with a name
lookup, rather than using Map directly. But I'll do that in
a separate commit.