Add a test that join is equivalent to a select on a cartesian product.
In the process, I had to add the 'constantExpression' and
'constantCondition' generators, for generators that don't
include any field references. This allows 'conditionAndSatisfingTuples'
to handle an empty list of attribute names.

There is too much code duplication between 'condition' and
'constantCondition', and between 'expression' and 'constantExpression'.
I'll tackle that in a separate clean-up commit.