Get rid of the 'constantExpression' and 'constantCondition' generators.
It's possible to discard the 'constantExpression' and 'constantCondition'
generators by making the types of 'condition' and 'expression' a little
more general, by making the name generator optional.

For example, instead of

 expression :: (...) => Gen n -> Int -> Gen (Expression n d m)

we generalize to

 expression :: (...) => Maybe (Gen n) -> Int -> Gen (Expression n d m)

and enable the attribute reference generator only when there really
is a name generator. Doing this cuts the code in half and permits
a simplification in 'conditionAndSatisfyingTuples'.