Add the Relational.RelVars module.
I'm about to start some feature extensions and some refactoring.
In preparation for implementing a simple in-memory database,
I'm adding a type for relational variable names. A 'database'
will contain named relational variables, associated with
relational values. This module isn't used anywhere yet, but
it will be in a few commits.