Add the Relational.Error module.
Currently, the Relational class permits an unnecessarily wide
range of monads. For example, any error monad is permitted.
This isn't what we want, since different kinds of relations
may need to be closely associated with particular kinds of
monads. Further, there are situations where we may want
to be able to do IO in certain implementations -- ones backed
by a real database, for example, or file system storage of
some kind.

If we're going to restrict the monads, we might as well
have a standard scheme for errors too. Here is the current
error type. In the next couple of commits, I'll require
that the monads be instances of (MonadError (RelationalError n)).