Replace the 'Relational' class with 'MonadRelational'.
The original 'Relational' class had methods that were supposed
to work any error monad. This isn't what we want.

In general, there is a close relationship between the
representation of relations and the monad in which relational
computations can be done. For example, if the representation
of relations requires doing disk I/O or talking to a database
over a socket, then the monad must be capable of doing I/O.

Therefore, before I make implementation changes, I've decided
to change class

 Relational n d r | r -> n, r -> d

to the class

 MonadRelational n d r m | m -> r, r -> n, r -> d

so that the monad is one of the parameters. This also, I think,
makes the approach here similar to what's more typically done
in code involving monads.

This is a straightforward change that requires just a few
implementation changes, but, unfortunately, requires a lot
of type changes in client modules, particularly in the
unit tests.

Some highlights:

1. The Relational.Naive module exports RelationalMonad and
   the monad transformer RelationalMonadT.

2. In RelationalProps, I had to change the 'prop_X' functions
   so they were less polymorphic, using RelationalMonad.
   That's in fact how they are used in all the tests, but
   in a way it's a pity.

3. In ConditionGen and some other places, I now use RelationalMonad
   in places that before used (Either String)

While making the changes, I realized it would have been (much)
shorter to just use the composition

 noErr . propM_X

in the QuickCheck tests directly, instead of defining
the prop_X functions. I'll consider making that change
in order to shorten the unit test code.