Fix a silly unit test bug, and the five broken tests it reveals.
While changing 'Relational' into 'MonadRelational', I noticed that
the 'noErr' function in RelationalProps has a glaring error in it.
This function converts the monadic propM_X functions into pure
prop_X functions, and was defined like so:

 noErr :: RelationalMonad d Bool -> Bool
 noErr = either (const False) (const True) . runRel

The left alternative is encountered when the RelationalMonad
version of the computation throws an error, so that case is fine.
However, the right alternative is handled with (const True),
which throws away the actual result of the RelationalMonad
version. This is wrong, since the result may in fact be
False, indicating that the property does not hold.

Sure enough, when I fixed this, I discovered that five of
the unit tests break. The tests also take much longer to
run now (I'll look into whether the test data is bigger
than needed later). Fortunately, all five failures
are the result of bugs in the tests themselves, rather
than bugs in RelationalMonad. Too bad I've been kidding
myself with these tests for the last three months :-/

The five bad tests are:

1. propM_renameIsReversible was comparing the wrong
   two relations

2. propM_canRemoveIntermediateRenames was also comparing
   the wrong two relations

3. propM_intersectionWithEmptyIsEmpty was also comparing
   the wrong two relations. It was in fact testing
   the obviously wrong condition

     r == r `intersect` empty

   instead of

     empty == r `intersect` empty

4. propM_projectionCommutesWithDifference is simply
   an incorrect test, since projection does not in
   fact commute with difference. Here is a counter-example:

     X = { A,B | [1,1], [2,2] }
     Y = { A,B | [1,2] }

   The relations X and Y are disjoint, but their projections
   on 'A' are not. So

     p[A] (X - Y) = p[A] X = {A | [1], [2]}

   whereas

     (p[A] X) - (p[A] Y) = {A | [1], [2]} - {A | [1]}
                         = {A | [2]}

5. propM_projectionCommutesWithIntersection is also
   an incorrect assertion. Consider relations X and Y
   again from (4). Since X and Y are disjoint, their
   intersection and any projection of it are empty:

     p[A] (X & Y) = p[A] {A,B|} = {A|}

   But

     (p[A] X) & (p[A] Y) = {A | [1],[2]} & {A | [1]}
                         = {A | [1]}

   which is not empty.

I've corrected (1), (2) and (3), and simply removed (4) and (5).