Add the Relational.ColName module, defining column names.
The MonadRelational class is arguably too general. While there
are reasons for allowing the type of data to vary, it's probably
not necessary to allow the type of column names to vary. Column
names, even if they have to be automatically generated, can
almost certainly be character strings.

So, as with RelNames, let's add a ColName type that wraps
a Text value.

This commit just adds the module. In the next commit, I'll
simplify the MonadRelational class so the type of column names
is not a variable, but fixed to ColName. That commit will
require changing Conditions and some of the unit test code,
but the changes should all be in the types.