Fix the type of column or attribute names to ColName.
This commit removes the type variable for attribute names
from the Expression, Condition, RelationalError, and
MonadRelational types, and makes several necessary changes
in Relational.Naive and related modules, as well as the
unit tests:

- Relational.Naive.Signature now uses Relational.ColName
  instead of Relational.Naive.AttrName

- we simplify a couple of function types in
  Relational.Naive.Signature by using ColName instead of
  an instance of the type class FromAttrName. There is
  no corresponding class for ColName, and I don't think
  the class really adds anything.

- we're able to remove some explicit type
  annotations that were required at uses of
  Relational.Naive.Signature.toList

- the AttrNameGen module in the unit tests becomes
  ColNameGen

- the types of many top-level functions in the unit tests
  change, either by eliminating a type variable for names,
  or by replacing that type variable with ColName

The changes are numerous, but the vast majority of them
are just straightforward changes in types.

At this point, Relational.Naive.AttrName is unused. I'll
remove it in the next commit.