Add the MonadDB class.
To create a naive in-memory database, or to front external
databases, we need to include relational variables, and operations
to create, destroy, read, and update them.

I've decided to add a second class, MonadDB, which extends
MonadRelational with these kinds of operations. I could have
simply extended MonadRelation, but I'd like to have the option
of having read-only relations -- say as a query interface to
some sort of other data.

Now that there are MonadRelational and MonadDB type classes,
I've also edited the Haddock docs slightly.