Add Relational.Naive.evalPure.
In the next few commits, I'll be changing the internal structure
of RelationalMonad and RelationalMonadT. However, there are some
test modules that rely on the current structure of RelationalMonad,
using 'runRel' to turn a value in RelationalMonad into a value
in 'Either RelationalError'. Instead, I've added a function

 evalPure :: RelationalMonad d a -> Either RelationalError a

which hides the details of the conversion, and switched to
using it in the dependent test modules.