Allow running unit tests with profiling and Haskell Program Coverage.
The unit tests are rather slow, and I suspect that the test relations
are bigger than they need to be. In order to confirm that, we'll need
to know where all the time is being spent, so I've added a 'proftests'
target that turns on profiling for the unit tests.

Since it's also useful to know how much of the code the tests are
executing, I've also added a 'covtests' target that adds HPC support.