Slightly improve the use of size in generating relations.
I noticed with 'sample' that arbitrary relations sometimes have
very large signatures and very few tuples. It turned out that
the 'Arbitrary' declaration for relations was generating
arbitrary signatures with the number of tuples varying between
0 and 6!

That's almost certainly not what I had intended. Varying
the signature length between 0 and 6 is likely what I meant.
Since arbitrary list generation is sensitive to the implicit
size, I've driven the max signature size of six into the
'relationOfSize' function. Also, it's sufficient to generate
signatures with

 resize 6 arbitrary

We're probably still generating relations that are bigger
than they need to be. Profiling the tests shows that about
a quarter of the time is spent in 'relUnsafeAddTuple' in
the Relational.Naive.Relation module. The implementation
of relations is *very* naive, and 'relUnsafeAddTuple'
could be quite a bit more efficient than it is, but I think
the relations could be smaller. I might experiment with
that a little more in the next couple of commits.

I've also given the 'tuples' function a parameter for the
data generator, so it can be used in 'relationWithSig'
and 'relationWithSigAndSize' and a couple of other functions.