Reduce the size of relations in generators that make multiple relations.
We've got a few heavily-used generators for producing pairs, triples,
and quadruples of relations:

 unionCompatiblePair
 unionCompatibleTriple
 unionCompatibleFour
 productCompatiblePair

When the implicit size is 's', make the relations of size 's/n' where
'n' is the tuple size. These changes further reduce the time to
run the current batch of unit tests to around 17 seconds on my
laptop. At this point, a quarter of the time is spent in
ConditionGen.satisfiableCondition, so further performance improvements
should concentrate on that generator.