Do join tests on relations over booleans.
The test on 'prop_joinLikeSelectOnProd' was being done on relations
over 'Small'. 'Small' is a five-element enumeration, but the
'productCompatiblePairAndSatisfiableCondition' generator would
generate two relations of up to three columns each. In the case
that there are six columns, this means the set of all tuples
has 5 ^ 6 = 15625 elements. It seems that this is what was
making 'satisfiableCondition' run so slowly.

At first I tried making 'satisfiableCondition' a bit smarter
by generating Conditions that tried to avoid being obviously
false (e.g. CondFalse, a conjunction containing CondFalse,
and so on). This had almost no effect on the running time.

Then I switched to using relations over Bool. With the maximum
of six columns in the join, there are 2 ^ 6 = 64 possible
tuples, and this really speeds the tests up. They're now down
to a little more than eight seconds.

I think the problem is that the generator for product-compatible
pairs and a satisfiable condition forces the evaluation
of the full list of satisfying tuples produced by
'satisfiableCondition', in order to pick a sub-list of the
tuples for inclusion in the generated relations. With 15625
tuples, this is quite expensive.

Now we're back to 'relUnsafeAddTuple' being the dominant
contributor to the running time. But the tests are fast enough
that there is no huge hurry to improve the performance of
naive relations just for the tests.