Remove unused code in the unit tests.
Looking at the marked up sources produced by 'hpc markup',
I noticed that two functions in the unit tests are not used
at all:

 ConditionGen.attrNamesIn
 SignatureGen.signatureAndNewName

I think these were part of approaches to tests that I tried
and then abandoned. There is no point in having this stuff
around, so we might as well delete it.