Fix a bug in the join tests: we never added arbitrary tuples to relations.
Code coverage markup of the RelationGen function
'productCompatiblePairAndSatisfiableCondition' showed that the
generated relations would never include arbitrary additional
tuples. In the tuples' function, in the expression

 tuples arbitrary (Sig.size sig)

the code coverage markup showed that the expressions 'arbitrary'
and '(Sig.size sig)' were never executed. Looking at the argument
to 'resize', we can see why:

 min 0 (n - length common)

If n (the current size) is greater than zero, we always resize
to zero. We should be using 'max' instead of 'min'. I also noticed
that earlier in the generator body we pick up to the current size
tuples from the satisfying tuples. The test is more interesting
if we limit the satisfying tuples to half the available size for
the relations, and then fill the rest with arbitrary tuples.

With these changes, code coverage now shows that all the code
executes in 'productCompatiblePairAndSatifiableCondition'.

I hadn't realized until now that the code coverage report on
the unit test code can be useful -- it can reveal places where
test code does not in fact run.