Remove the one remaining use of Relation.Naive.Error.
I didn't notice in the last commit that Relational.Naive.Error.die
was still used in Relational.Naive.Signature.safeFromList. Change
'safeFromList' so it insists on a MonadError MonadRelational and
uses the 'duplicatedColName' error function.

At this point, I've removed Relational.Naive.Error from the
exposed modules in the cabal configuration file, and I've confirmed
that Relational.Naive.Error is no longer compiled in -- it has
disappeared from the code coverage report in the HPC-instrumented
unit tests. I'll remove the module entirely in the next commit.