Add Relational.Naive.RelVars and some associated unit tests.
In order to implement the new MonadDB, we'll need to build a
state monad that tracks the current set of relational variables.
This commit adds (part of) the state, the RelVars type, which
is a mapping from RelNames to Relations. This uses Data.Map
under the covers, but hides most of its interface.