Make things work with ghc 7.0.3, Haskell Platform 2011.2.0.1.
In order to get the main package to compile, I had to relax the
type constraint on mtl to allow mtl 2.

I then noticed that compilation in the unit tests was broken:
the ConditionGen module was unhappy about the type environment
constraint

  CoArbitrary [d]

in the type signatures for the functions

  expression
  condition

Now, I *swear* these tests were working when I last made
commits almost eight months ago. In fact, the second-last
commit,

18391f8e36c2de18169841803e6f4a8d8055d154 Remove Relational.Naive.Error.

made on December 30, 2010 says


  ... and the unit tests still run successfully.

Investigating the problem, it seems that now GHC wants the
constraint

  CoArbitrary d

on 'condition' and 'expression'. Now, these two functions
need the CoArbitrary constraint because they invoke
'noFailFunction':

 noFailFunction :: (CoArbitrary [d], Arbitrary r, Monad m) =>
                   Gen ([d] -> m r)
 noFailFunction = arbitrary >>= \f -> return (return . f)

This is used to generate functions used in the CondCall and
ExpCall variants of Condition and Expression. I remember that
I deliberately relaxed the constraint in the type environment
for 'noFailFunction' from

  CoArbitrary d

to

  CoArbitrary [d]

since this is the weaker constraint. Now, the Test.QuickCheck
module does provide an instance declaration for (CoArbitrary [a])
given an instance declaration for (CoArbitrary a). In practice,
we'd never declare a one-of instance (CoArbitrary [d]) for some
particular type 'd'. But I figured I'd stick with the principle
of using the most relaxed constraint possible.

But with GHC 7.0.3, this does not compile. GHC insists that the
required constraint is (CoArbitrary d). Everything compiles now
and the tests run, but I wonder what has changed. I can't remember
whether I was using an earlier version of GHC 7 or GHC 6.12 last
December. I'll have to look through the GHC release notes and
see if there is an explanation of what I'm doing wrong.